+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   - `assertNotEquals` -> `assert_not_equals`
   - `assertEmpty` -> `assert_empty`
   - `assertNotEmpty` -> `assert_not_empty`
+  - `assertContains` -> `assert_contains`
 
 ## [0.6.0](https://github.com/TypedDevs/bashunit/compare/0.5.0...0.6.0) - 2023-09-19